switch to spring boot 3 / jdk 17 baseline (#277)

+++ ci/validate-cas-mgmt-overlay.sh
@@ -2,8 +2,13 @@
 
 source ./ci/functions.sh
 
+FETCH_OVERLAY="false"
 while (( "$#" )); do
     case "$1" in
+    --fetch-only)
+        FETCH_OVERLAY="true"
+        shift 1
+        ;;
     --cas)
         CAS_VERSION="$2"
         shift 2
@@ -19,18 +24,24 @@ while (( "$#" )); do
     esac
 done
 
-java -jar app/build/libs/app.jar &
-pid=$!
-sleep 30
-rm -Rf tmp &> /dev/null
-mkdir tmp
-cd tmp
-curl http://localhost:8080/starter.tgz --connect-timeout 60 \
-    -d casVersion=${CAS_VERSION} -d bootVersion=${BOOT_VERSION} \
-    -d type=cas-management-overlay | tar -xzvf -
-kill -9 $pid
+mkdir -p tmp
+cd tmp || exit
+echo "Working directory: ${PWD}"
 
-echo "Building CAS Mgmt Overlay"
+if [[ "${FETCH_OVERLAY}" == "true" ]]; then
+  java -jar ../app/build/libs/app.jar &
+  pid=$!
+  sleep 30
+  curl http://localhost:8080/starter.tgz --connect-timeout 60 \
+      -d casVersion=${CAS_VERSION} -d bootVersion=${BOOT_VERSION} \
+      -d type=cas-management-overlay | tar -xzvf -
+  kill -9 $pid
+  ls
+  printgreen "Downloaded CAS overlay ${CAS_VERSION} successfully"
+  exit 0
+fi
+
+echo "Building CAS Mgmt Overlay in ${PWD}"
 ./gradlew clean build --no-daemon
 
 dname="${dname:-CN=cas.example.org,OU=Example,OU=Org,C=US}"
